refactor: lib file structure (#61)
* refactor: improve lib file structure

* chore: format code

* chore: refactor constants & interfaces file location

* refactor: move enums file location

* fix: missing interface I prefix
