+++ hooks/index.ts
@@ -1,7 +1,7 @@
 export { default as useForm } from './useForm'
-export { default as useUploadFiles } from './useUploadFiles'
-export { default as useRPC } from './useRPC'
-export { default as useUploadCache } from './useUploadCache'
-export { default as useSearchBar } from './useSearchBar'
-export { default as useVerifyCandyMachineV2 } from './useVerifyCandyMachineV2'
 export { default as useRemoveCandyMachineAccount } from './useRemoveCandyMachineAccount'
+export { default as useRPC } from './useRPC'
+export { default as useSearchBar } from './useSearchBar'
+export { default as useUploadCache } from './useUploadCache'
+export { default as useUploadFiles } from './useUploadFiles'
+export { default as useVerifyCandyMachineV2 } from './useVerifyCandyMachineV2'
