+++ lib/candy-machine/index.ts
@@ -1,0 +1,2 @@
+export * from './update'
+export * from './upload'
