+++ lib/candy-machine/types.ts
@@ -1,0 +1,1 @@
+export type Account = string | string[] | undefined
