+++ lib/candy-machine/update.ts
@@ -1,7 +1,6 @@
-import { PublicKey } from '@solana/web3.js'
 import { Program } from '@project-serum/anchor'
-import { saveCache } from '../cache'
-import { Cache } from 'lib/interfaces'
+import { PublicKey } from '@solana/web3.js'
+import { ICache, saveCache } from 'lib/cache'
 
 export async function updateV2({
     newSettings,
@@ -21,7 +20,7 @@ export async function updateV2({
     newAuthority: string
 }) {
     try {
-        const cacheContent: Cache = JSON.parse(cache)
+        const cacheContent: ICache = JSON.parse(cache)
         const env = cacheContent.env
         const cacheName = cacheContent.cacheName
