+++ lib/candy-machine/upload/arweave.ts
@@ -1,10 +1,10 @@
-import * as anchor from '@project-serum/anchor'
 import { calculate } from '@metaplex/arweave-cost'
-import { ARWEAVE_PAYMENT_WALLET, ARWEAVE_UPLOAD_ENDPOINT } from '../constants'
-import { sendTransactionWithRetryWithKeypair } from './transactions'
-import { Manifest } from '../types'
-import { getFileExtension } from './helpers'
+import * as anchor from '@project-serum/anchor'
 import { AnchorWallet } from '@solana/wallet-adapter-react'
+import { ARWEAVE_PAYMENT_WALLET, ARWEAVE_UPLOAD_ENDPOINT } from 'lib/candy-machine/constants'
+import { Manifest } from 'lib/types'
+import { getFileExtension } from './helpers'
+import { sendTransactionWithRetryWithKeypair } from './transactions'
 
 /**
  * @param fileSizes - array of file sizes
